Provide backwards compatibility in rotate extrude.
This introduces a new axis parameter to the rotate_extrude module
in the future this can be used to rotate around a given axis.
when the parameter is not set it will use compatibility mode which means
rotating the xy plane around the z axis as if the xy plane was actually
the xz plane.